Followup to b4f56a9d: Followup to  Ensure parse.php --linting emits output
* This was one of the reasons why $env->pageWithOldid check had
  been pushed into extension/src/Config/DataAccess.php previously.

* But a better solution is to set the pageWithOldid flag always when
  parsing on the commandline.

* Update toolcheck script to ensure this doesn't fail in the future.

+++ bin/toolcheck.php.sh
@@ -39,5 +39,15 @@ php $BIN/parse.php --pboutfile $TMPPB < $TMPWT |
 php $BIN/parse.php --pbinfile $TMPPB --selser \
     --oldtextfile $TMPWT --oldhtmlfile $TMPORIG < $TMPEDIT
 
+# Linting
+echo "<div>foo" | php $BIN/parse.php --linting
+
+# Ensure lint output (exactly 1 line) is present!
+x=`echo "<div>foo" | php $BIN/parse.php --linting 2>&1 > /dev/null | wc -l`
+if [ ! $x -eq 1 ]
+then
+	exit -1
+fi
+
 # clean up
 /bin/rm $TMPWT $TMPORIG $TMPEDIT $TMPPB
